requirements.txt
# PyTorch with MPS support for Mac M-series chips
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
numpy>=1.21.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Web scraping and progress bar
requests>=2.28.0
tqdm>=4.64.0

# Optional: For additional image formats
opencv-python>=4.5.0

# FastAPI and web server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
psutil>=5.9.0